test: unify hook registry tests into test_entry_points.py
Moves the new tests for `add_hook` and `get_hooks` into
`tests/unit/digest/test_entry_points.py`, as requested. Includes
a shared fixture to correctly reset `_HOOKS` and `_EP_HOOKS`
after each test. Deleted `test_hooks.py`.
